Added Room class
Room will hold the details of each room, right now being the housekeeping status, size of bed, type of room, quality of the room, if smoking is permitted (possibly pay for it to be allowed?) and a room number.  This class also has constants that store the relevant max and min of room numbers on each floor.