Added Reservations class
This class is not complete by any means, still a work in progress. The Reservations is intended to handle like a controller would directing commands from the user interface to the necessary classes and methods. The Reservations primary goal will be to encrypt IDs like roomIDs, reservationIDs, userIDs, and any future ID requiremens. This encryption will ensure guest privacy, and allow for the simplification of performing commands if the ID is known because the ID will hold all  possible relevant information.